requirements.txt
appdirs==1.4.4
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
coala-utils==0.5.1
cssselect==1.1.0
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
facebook-scraper==0.1.12
fake-useragent==0.1.11
idna==2.9
importlib-metadata==1.6.0
instagram-explore==0.3.0
lxml==4.5.1
Markdown==3.2.2
parse==1.15.0
pyee==7.0.2
PyJWT==1.7.1
pyppeteer==0.2.2
pyquery==1.4.1
pytz==2020.1
requests==2.23.0
requests-html==0.10.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tqdm==4.46.1
twitterscraper==1.4.0
urllib3==1.25.9
w3lib==1.22.0
websockets==8.1
zipp==3.1.0
